requirements.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
feedparser==6.0.11
Flask==3.0.3
Flask-Mail==0.10.0
fonttools==4.58.0
fqdn==1.5.1
future==1.0.0
fuzzywuzzy==0.18.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gnewsclient==1.12
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.8
Levenshtein==0.26.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
newsapi-python==0.2.7
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pillow==11.2.1
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-Levenshtein==0.26.1
pytz==2025.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
RapidFuzz==3.10.1
ratelim==0.1.6
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
scipy==1.15.3
Send2Trash==1.8.2
setuptools==69.1.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
threadpoolctl==3.6.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.14
tzdata==2025.1
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.4
widgetsnbextension==4.0.9
